+++ test/tests.jl
@@ -17,17 +17,16 @@ using Clustering            # test/MSA/Hobohm.jl
 using NamedArrays           # array
 using StatsBase             # WeightVec
 using PairwiseListMatrices  # getlist
-using HTTP                  # HTTP.Exceptions.StatusError
 
 const DATA = joinpath(@__DIR__, "data")
 
 # Utils
-@testset "Utils" begin
+@testset verbose=true "Utils" begin
     include("Utils/GeneralUtils.jl")
 end
 
 # MSA
-@testset "MSA" begin
+@testset verbose=true "MSA" begin
     include("MSA/Residues.jl")
     include("MSA/Alphabet.jl")
     include("MSA/ThreeLetters.jl")
@@ -47,7 +46,7 @@ end
 end
 
 # Information
-@testset "Information" begin
+@testset verbose=true "Information" begin
     include("Information/ContingencyTables.jl")
     include("Information/Counters.jl")
     include("Information/InformationMeasures.jl")
@@ -58,7 +57,7 @@ end
 end
 
 # PDB
-@testset "PDB" begin
+@testset verbose=true "PDB" begin
     include("PDB/PDB.jl")
     include("PDB/Contacts.jl")
     include("PDB/Kabsch.jl")
@@ -68,17 +67,17 @@ end
 end
 
 # SIFTS
-@testset "SIFTS" begin
+@testset verbose=true "SIFTS" begin
     include("SIFTS/SIFTS.jl")
 end
 
 # Pfam
-@testset "Pfam" begin
+@testset verbose=true "Pfam" begin
     include("Pfam/Pfam.jl")
 end
 
 # Scripts
-@testset "Scripts" begin
+@testset verbose=true "Scripts" begin
     include("Scripts/Template.jl")
     include("Scripts/Scripts.jl")
 end